version avec les sweet alert fonctionnant

+++ src/app/liste-auteurs/liste-auteurs.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AuteursService } from '../services/auteurs.service';
 import { Auteurs } from '../models/Auteurs';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-liste-auteurs',
@@ -8,23 +9,48 @@ import { Auteurs } from '../models/Auteurs';
   styleUrls: ['./liste-auteurs.component.css']
 })
 export class ListeAuteursComponent implements OnInit {
-  listAuteurs: Auteurs[]=[];
+  listAuteurs: Auteurs[] = [];
 
-  constructor(private AuteursService:AuteursService) { }
+  constructor(private AuteursService: AuteursService) { }
 
   ngOnInit() {
     this.AuteursService.getAll().subscribe(
-      data=>{
+      data => {
         console.log(this.listAuteurs)
-        this.listAuteurs=data;
+        this.listAuteurs = data;
       }
     );
   }
-supprimerAuteurs(id:number){
-  this.AuteursService.deleteOne(id).subscribe(
-    data=>{
-      this.ngOnInit();
-    }
-  )
-}
+  supprimerAuteurs(id: number) {
+    Swal.fire({
+      icon: 'error',
+      title: 'Desolé...',
+      text: 'Cet auteur ne peux etre supprimé!',
+      footer: '<a href>Why do I have this issue?</a>'
+    })
+    Swal.fire({
+      title: 'Etes vous sûr?',
+      text: "Vous ne pourrez pas revenir en arrière!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Oui, continuer!'
+    }).then((result) => {
+      if (result.value) {
+        this.AuteursService.deleteOne(id).subscribe(
+          data => {
+            Swal.fire(
+              'Supprimé!',
+              'Le fichier a été supprimé.',
+              'success'
+            )
+            this.ngOnInit();
+          }
+        );
+
+      }
+    })
+
+  }
 }
