+++ src/app/liste-categorie/liste-categorie.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Categorie } from '../models/Categorie';
 import { CategorieService } from '../services/categorie.service';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-liste-categorie',
@@ -20,11 +21,35 @@ listCategorie: Categorie[]=[];
   );
   }
   SupprimerCategorie(id:number){
-   this.CategorieService.deleteOne(id).subscribe(
-     data=>{
-       this.ngOnInit();
+    Swal.fire({
+      icon: 'error',
+      title: 'Desolé...',
+      text: 'Cet auteur ne peux etre supprimé!',
+      footer: '<a href>Why do I have this issue?</a>'
+    })
+    Swal.fire({
+      title: 'Etes vous sûr?',
+      text: "Vous ne pourrez pas revenir en arrière!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Oui, continuer!'
+    }).then((result) => {
+      if (result.value) {
+        this.CategorieService.deleteOne(id).subscribe(
+          data => {
+            Swal.fire(
+              'Supprimé!',
+              'Le fichier a été supprimé.',
+              'success'
+            )
+            this.ngOnInit();
+          }
+        );
 
-     }
-   );
+      }
+    })
+   
   }
 }
